Fix Linux problems: mouse sensitivity, accept external LD_LIBRARY_PATH (steam runtime, etc), ignore GLFW minimize error

+++ 3dEngine/src/scene/Renderer.h
@@ -11,7 +11,7 @@ namespace urchin {
             virtual bool onKeyPress(unsigned int) = 0;
             virtual bool onKeyRelease(unsigned int) = 0;
             virtual bool onChar(char32_t) = 0;
-            virtual bool onMouseMove(int, int) = 0;
+            virtual bool onMouseMove(double, double) = 0;
             virtual void onDisable() = 0;
 
             virtual void prepareRendering(float) = 0;
