+++ 3dEngine/src/scene/renderer3d/Renderer3d.h
@@ -75,7 +75,7 @@ namespace urchin {
             bool onKeyPress(unsigned int) override;
             bool onKeyRelease(unsigned int) override;
             bool onChar(char32_t) override;
-            bool onMouseMove(int, int) override;
+            bool onMouseMove(double, double) override;
             void onDisable() override;
 
             //scene
